Changes to AbortSignal handling retries
