+++ common/config/azure-pipelines/build-test.yml
@@ -7,7 +7,7 @@ parameters:
   default: true
 
 variables:
-  node18Version: '18.13.0'
+  node20Version: '20.18.2'
 
 trigger: none
 
@@ -15,13 +15,13 @@ jobs:
   - job: BuildAndTest
     strategy:
       matrix:
-        linux-node-18:
+        linux-node-20:
           imageName: 'ubuntu-latest'
-          nodeVersion: $(node18Version)
+          nodeVersion: $(node20Version)
           runFrontendIntegrationTests: true
-        windows-node-18:
+        windows-node-20:
           imageName: 'windows-latest'
-          nodeVersion: $(node18Version)
+          nodeVersion: $(node20Version)
           runFrontendIntegrationTests: true
     pool:
         vmImage: $(imageName)
