+++ common/config/azure-pipelines/templates/publish.yml
@@ -1,8 +1,8 @@
 steps:
   - task: NodeTool@0
-    displayName: Install Node@$(node18Version)
+    displayName: Install Node@$(node20Version)
     inputs:
-      versionSpec: $(node18Version)
+      versionSpec: $(node20Version)
       checkLatest: true
 
   - script: node common/scripts/install-run-rush.js install --purge
