+++ storage/core/src/common/index.ts
@@ -2,6 +2,7 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
+export * from "./AbortSignal";
 export * from "./Types";
 export * from "./Interfaces";
 export * from "./Utils";
