+++ storage/google/package.json
@@ -49,7 +49,7 @@
     "@google-cloud/storage-control": "^0.5.0",
     "@itwin/cloud-agnostic-core": "workspace:*",
     "@itwin/object-storage-core": "workspace:*",
-    "axios": "^1.10.0",
+    "axios": "^1.12.2",
     "google-auth-library": "^10.1.0"
   },
   "devDependencies": {
@@ -60,7 +60,7 @@
     "@types/chai": "^4.3.8",
     "@types/chai-as-promised": "^7.1.2",
     "@types/mocha": "^10.0.1",
-    "@types/node": "^18.11.18",
+    "@types/node": "^20.17.20",
     "@types/sinon": "^10.0.11",
     "chai": "^4.3.10",
     "chai-as-promised": "^7.1.1",
