+++ tests/backend-storage-unit/package.json
@@ -49,7 +49,7 @@
     "@types/chai": "^4.3.8",
     "@types/chai-as-promised": "^7.1.2",
     "@types/mocha": "^10.0.1",
-    "@types/node": "^18.11.18",
+    "@types/node": "^20.17.20",
     "cspell": "^8.16.0",
     "eslint": "^8.57.1",
     "rimraf": "^2.6.2",
