+++ tests/backend-storage/package.json
@@ -54,7 +54,7 @@
     "@types/chai-as-promised": "^7.1.2",
     "@types/fs-extra": "^9.0.13",
     "@types/mocha": "^10.0.1",
-    "@types/node": "^18.11.18",
+    "@types/node": "^20.17.20",
     "@types/yargs": "^15.0.5",
     "cspell": "^8.16.0",
     "eslint": "^8.57.1",
