+++ tests/backend-storage/src/common-tests/ClientStorage.test.ts
@@ -189,6 +189,7 @@ describe(`${ClientStorage.name}: ${clientStorage.constructor.name}`, () => {
         const downloadUrl = await serverStorage.getDownloadUrl(uploadedFile);
         const abortController = new AbortController();
 
+        abortController.abort();
         const downloadPromise = clientStorage.download({
           storageType: storageType,
           url: downloadUrl,
@@ -196,8 +197,6 @@ describe(`${ClientStorage.name}: ${clientStorage.constructor.name}`, () => {
           localPath: `${testDownloadPath}/download-url.txt`,
           abortSignal: abortController.signal,
         });
-
-        abortController.abort();
 
         let wasAborted = false;
         try {
@@ -530,6 +529,7 @@ describe(`${ClientStorage.name}: ${clientStorage.constructor.name}`, () => {
             );
           const abortController = new AbortController();
 
+          abortController.abort();
           const downloadPromise = clientStorage.download({
             reference: uploadedFile,
             transferConfig: downloadConfig,
@@ -537,8 +537,6 @@ describe(`${ClientStorage.name}: ${clientStorage.constructor.name}`, () => {
             localPath: path.join(testDownloadPath, "download-config.txt"),
             abortSignal: abortController.signal,
           });
-
-          abortController.abort();
 
           let wasAborted = false;
           try {
