+++ tests/frontend-storage/package.json
@@ -46,8 +46,8 @@
   "devDependencies": {
     "@itwin/object-storage-common-config": "workspace:*",
     "@types/express": "^4.17.13",
-    "@types/node": "^18.11.18",
-    "axios": "^1.10.0",
+    "@types/node": "^20.17.20",
+    "axios": "^1.12.2",
     "cspell": "^8.16.0",
     "eslint": "^8.57.1",
     "express": "^4.18.0",
